+++ src/GIntervalsBigSet2D.h
@@ -62,8 +62,8 @@ public:
 	virtual GIntervals2D::const_iterator get_chrom_begin() const { return m_intervals.begin(); }
 	virtual GIntervals2D::const_iterator get_chrom_end() const { return m_intervals.end(); }
 
-	virtual size_t iter_index() const { return m_iter_index; }
-	virtual size_t iter_chrom_index() const { return m_iter_chrom_index; }
+	virtual uint64_t iter_index() const { return m_iter_index; }
+	virtual uint64_t iter_chrom_index() const { return m_iter_chrom_index; }
 
 	virtual const GInterval2D &cur_interval() const { return *m_iinterval; }
 
@@ -76,8 +76,8 @@ private:
 	GIntervals2D::const_iterator m_iinterval;
 	int                          m_cur_chromid;
 	int                          m_iter_chromid;
-	size_t                       m_iter_index;
-	size_t                       m_iter_chrom_index;
+	uint64_t                       m_iter_index;
+	uint64_t                       m_iter_chrom_index;
 	Compare_t                    m_compare;
 	bool                         m_do_sort;
