+++ src/GenomeComputeStrandAutocorr.cpp
@@ -107,8 +107,8 @@ SEXP gcompute_strands_autocorr(SEXP _infile, SEXP _chrom, SEXP _binsize, SEXP _m
 		if (max_coord < 0 || max_coord > chromsize)
 			max_coord = chromsize;
 
-		vector<int> forward((size_t)ceil(chromsize / binsize), 0);
-		vector<int> reverse((size_t)ceil(chromsize / binsize), 0);
+		vector<int> forward((uint64_t)ceil(chromsize / binsize), 0);
+		vector<int> reverse((uint64_t)ceil(chromsize / binsize), 0);
 
 		BufferedFile infile;
 		infile.open(infilename, "r");
@@ -122,8 +122,8 @@ SEXP gcompute_strands_autocorr(SEXP _infile, SEXP _chrom, SEXP _binsize, SEXP _m
 		string str[NUM_COLS];
 		int min_off = (int)(-maxread / binsize);
 		int max_off = (int)(maxread / binsize);
-		size_t min_idx = (size_t)(max_off + min_coord / binsize);
-		size_t max_idx = (size_t)(max_coord / binsize - max_off - 1);
+		uint64_t min_idx = (uint64_t)(max_off + min_coord / binsize);
+		uint64_t max_idx = (uint64_t)(max_coord / binsize - max_off - 1);
 
 		if (min_idx >= forward.size() || (int64_t)max_idx < 0)
 			verror("Not enough data to calculate auto correlation.");
@@ -167,11 +167,11 @@ SEXP gcompute_strands_autocorr(SEXP _infile, SEXP _chrom, SEXP _binsize, SEXP _m
 							break;
 
 						if (str[STRAND_COL] == "+" || str[STRAND_COL] == "F") {
-							size_t idx = (size_t)(coord / binsize);
+							uint64_t idx = (uint64_t)(coord / binsize);
 							forward[idx] = min(MAX_COV, forward[idx] + 1);
 						}
 						else if (str[STRAND_COL] == "-" || str[STRAND_COL] == "R") {
-							size_t idx = (size_t)((coord + str[SEQ_COL].size()) / binsize);
+							uint64_t idx = (uint64_t)((coord + str[SEQ_COL].size()) / binsize);
 							reverse[idx] = min(MAX_COV, reverse[idx] + 1);
 						}
 
@@ -213,7 +213,7 @@ SEXP gcompute_strands_autocorr(SEXP _infile, SEXP _chrom, SEXP _binsize, SEXP _m
 		int64_t tot_rr = 0;
 		vector<double> tot_fr(max_off - min_off);
 
-		for (size_t i = min_idx; i < max_idx; i++) {
+		for (uint64_t i = min_idx; i < max_idx; i++) {
 			int cur_fr = forward[i];
 			int cur_rv = reverse[i];
 			tot_f += cur_fr;
